perf(HelpPage): drop gray-matter from client bundle
